Fetch and set metadata for readme
Instead of fetching file from backend for readme, when uploading a file,
the readme metadata is set with files found inside the archive.

Using this way, the metadata is stored inside the real readme metadata
and follow the generic way how it should be implemented.

To retreive the readme, you know request all the metadata available,
which contains readme inside. In addition, any others metadata set will
be returned as well.

If anyone upload an flist with specific metadata set, this still works.