tools: add hub-cleanup early version
Introduce a new 'hub-cleanup' tool which would cleanup zdb according to
local files requirement.

The tool is not finished yet but provide some early feature: it can now
load a set of flist and extract all files chunks from it. It keep a
global hashtable of all chunks.

Later, this hashtable can be used to compare zdb content and build a map
of needed and non-needed keys, making some cleanup possible